requirements.txt
numpy
tensorflow
streamlit
streamlit-webrtc
pillow
opencv-python
onnxruntime